+++ src/main/kotlin/blackjack/domain/card/Shape.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class Shape {
