+++ src/main/kotlin/blackjack/domain/gameResult/PlayerGameResult.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.gameResult
 
 data class PlayerGameResult(
     val playerName: String,
