+++ src/main/kotlin/blackjack/domain/gameResult/TotalGameResult.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.gameResult
 
 data class TotalGameResult(
