+++ src/main/kotlin/blackjack/domain/player/Player.kt
@@ -1,4 +1,6 @@
-package blackjack.domain
+package blackjack.domain.player
+
+import blackjack.domain.card.Cards
 
 class Player(
     val name: PlayerName,
