+++ src/main/kotlin/blackjack/domain/player/PlayerName.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.player
 
 @JvmInline
 value class PlayerName(val value: String) {
